add PosProductAdapter

+++ app/src/main/java/com/ahmadabuhasan/skripsi/adapter/PosProductAdapter.java
@@ -55,8 +55,8 @@ public class PosProductAdapter extends RecyclerView.Adapter<PosProductAdapter.My
     @SuppressLint("SetTextI18n")
     @Override
     public void onBindViewHolder(@NonNull MyViewHolder holder, int position) {
-        //final int getStock;
-        final double getStock;
+        final int getStock;
+        //final double getStock;
         final DatabaseAccess databaseAccess = DatabaseAccess.getInstance(this.context);
 
         databaseAccess.open();
@@ -72,8 +72,8 @@ public class PosProductAdapter extends RecyclerView.Adapter<PosProductAdapter.My
         databaseAccess.open();
         String weight_unit_name = databaseAccess.getWeightUnitName(weight_unit_id);
 
-        //int getStock1 = Integer.parseInt(product_stock);
-        double getStock1 = Double.parseDouble(product_stock);
+        int getStock1 = Integer.parseInt(product_stock);
+        //double getStock1 = Double.parseDouble(product_stock);
         if (getStock1 > 5) {
             TextView textView = holder.textView_Stock;
             getStock = getStock1;
